Apply new changes;

+++ tests/LocalDriverTest.php
@@ -9,7 +9,6 @@ use RonasIT\AutoDoc\Exceptions\MissedProductionFilePathException;
 class LocalDriverTest extends TestCase
 {
     protected static LocalDriver $localDriverClass;
-    protected static string $baseFileName;
     protected static string $baseFile;
     protected static string $tmpDocumentationFilePath;
     protected static array $tmpData;
@@ -19,25 +18,20 @@ class LocalDriverTest extends TestCase
         parent::setUp();
 
         $documentationDirectory = config('auto-doc.drivers.local.directory');
-        if (!str_ends_with($documentationDirectory, DIRECTORY_SEPARATOR)) {
-            $documentationDirectory .= DIRECTORY_SEPARATOR;
-        }
 
-        self::$baseFileName ??= 'documentation';
-        self::$baseFile ??= storage_path($documentationDirectory . self::$baseFileName . '.json');
+        self::$baseFile ??= storage_path("{$documentationDirectory}/documentation.json");
         self::$tmpDocumentationFilePath ??= storage_path('temp_documentation.json');
 
         self::$tmpData ??= $this->getJsonFixture('tmp_data');
 
-        config(['auto-doc.drivers.local.base_file_name' => self::$baseFileName]);
+        config(['auto-doc.drivers.local.base_file_name' => 'documentation']);
 
         self::$localDriverClass ??= new LocalDriver();
     }
 
     public function testDirectoryEndsWithDirectorySeparator()
     {
-        config(['auto-doc.drivers.local.directory' => config('auto-doc.drivers.local.directory').DIRECTORY_SEPARATOR]);
-        config(['auto-doc.drivers.local.base_file_name' => self::$baseFileName]);
+        config(['auto-doc.drivers.local.directory' => 'documentations'.DIRECTORY_SEPARATOR]);
 
         $driver = new LocalDriver();
         $driver->saveTmpData(self::$tmpData);
@@ -100,8 +94,10 @@ class LocalDriverTest extends TestCase
 
     public function testSaveDataWhenDirectoryNotExists()
     {
-        $documentationDirectory = config('auto-doc.drivers.local.directory');
-        rmdir(storage_path($documentationDirectory));
+        $documentationDirectory = 'test_directory';
+        if (is_dir($documentationDirectory)) {
+            rmdir(storage_path($documentationDirectory));
+        }
 
         self::$localDriverClass->saveTmpData(self::$tmpData);
